Exam_Engine/requirements.txt
fastapi
uvicorn
redis
python-dotenv
pydantic
sqlalchemy
asyncpg
psycopg2-binary
elasticsearch
structlog
elastic-apm
